[mod] c++, big number, 'divide_', performance
-subtle modification to the determination of the initial estimate
for the divisor into the dividend / remainder. the original algorithm
would miss the optimum guess when the head of the divisor was greater
than the head of the remainder - the comparison was performed on
relative lengths only. although the 'miss' wasn't fatal, the ultimate
determination that the guess was invalid, and subsequent drawing down
of further dividend digits to allow the calculation to continue wasted
a lot of cycles
-revert to basing divisor at 10^0 digit plus / minus shift inline
with the reciprocal '1' as it's simpler to follow